fix: widen SegmentSpec.length to long (TODO #12)
The on-disk segment_spec.length field is uint32 — the FlatBuffer
schema is `length: uint32`, and the flatc-generated accessor returns
it as a long via `(long) bb.getInt(...) & 0xFFFFFFFFL`. Java was
narrowing that back to int in three places, silently truncating any
segment between 2 GB and 4 GB into a negative value:

  - PostscriptParser.convertFooter: `(int) s.length()` cast
  - VortexWriter.SegRef.len: int field plus `(int) (bytesWritten - offset)` cast
  - SegmentSpec.length itself: int field, propagating the truncation
    through all callers

The README has been advertising "4 GB per segment (uint32 length)" as
a Parquet differentiator, but anything past 2 GB rounded into a
negative offset and crashed downstream slicing.

Widen `SegmentSpec.length` and `SegRef.len` to long, drop both casts,
and propagate the type through ScanIterator and VortexInspector.

ScanIterator.readFlatStats no longer materialises the entire segment
via `seg.asByteBuffer()` — ByteBuffer caps at 2 GB, so a 3 GB segment
would throw before any stats peek ran. Slice the FlatBuffer tail off
the MemorySegment first (small slice, fits in a ByteBuffer trivially).

PostscriptParser.convertFooter is now package-private so a unit test
can verify the conversion without crafting a full Vortex file.
PostscriptParserBigSegmentTest builds a synthetic FlatBuffer Footer
whose single segment_spec advertises length = 0xC0000000 (3 GB) at
offset = 4 GB, then asserts SegmentSpec.length() returns the full
long, not a truncated negative int.

End-to-end multi-GB round-trip (write a real >2 GB file, scan it via
ScanIterator) is deferred — slow to run, lives under a future slow
profile per TODO #12.

+++ core/src/main/java/io/github/dfa1/vortex/core/SegmentSpec.java
@@ -1,9 +1,13 @@
 package io.github.dfa1.vortex.core;
 
 /// Byte range and properties of one data segment in the file.
+///
+/// {@code length} is {@code long} because the wire format encodes it as
+/// {@code uint32} (up to 4 GB per segment). Storing it as a Java {@code int}
+/// silently truncates segments larger than 2 GB into negative values.
 public record SegmentSpec(
 		long offset,
-		int length,
+		long length,
 		byte alignmentExponent,
 		CompressionScheme compression
 ) {
